Move CardPreview.tsx to CardPreview/CardPreview.tsx
